Crate light

This is SIMPLE’s Light and Solar calculation module. It is responsible for:

Calculating Incident Solar Radiation in each surface: Contrary to EnergyPlus (and probably other tools I am less familiar with), this module uses Daylight Coefficients for performing this simulation. This method was stolen from the daylighting simulation world, and has the advantage of being extremely robust, and therefore capable of handling complex geometries. Perhaps the main drawback is that—because the concept of Thermal Zone does not fit within Lighting calculations (it is quite artificial for radiation purposes, actually)—reporting the “Solar Heat Gains” in a zone needs significant post-processing.

Calculating view factors for Infrared calculations: This is in development… for now, it does nothing.

Daylighting Calculations: Because this module is based on ray-tracing, it can perform daylight calculations. It is unclear, however, which climate based daylight metrics to include and how… if you have any idea, let me know.

Re-exports

pub use model::SolarModel;

Modules

model

The main export of this module: A Simulation Model for calculating solar and lighting factors.

optical_info 🔒

solar_surface 🔒

Structs

IRViewFactorSet

A set of view factors as seen by a ThermalSurface.

OpticalInfo

Information about the solar radiation and other optical elements of the whole model.

Constants

PI

Well, Pi.

Type Definitions

Float

The kind of Floating point number used in the library… the "float" feature means it becomes f32 and f64 is used otherwise.