Struct light::solar_surface::SolarSurface

pub struct SolarSurface {
    points: Vec<Point3D>,
    pub normal: Vector3D,
    pub receives_sun_front: bool,
    pub receives_sun_back: bool,
}

Structure that can help calculate solar radiation

It contains the normal of the original Surface and the points randomly sampled in each surface.

Fields
points: Vec<Point3D>normal: Vector3Dreceives_sun_front: boolreceives_sun_back: bool
Implementations

impl SolarSurface

const DELTA: Float = 0.001f64

Offset for the starting point of the rays.

pub fn new(
nrays: usize,
polygon: &Polygon3D,
receives_sun_front: bool,
receives_sun_back: bool
) -> Self

Creates a new Solar Surface

pub(crate) fn boundary_receives_sun(boundary: Result<&Boundary, String>) -> bool

pub fn calc_solar_dc_matrix(
list: &[SolarSurface],
scene: &Scene,
dc_factory: &DCFactory,
front_side: bool
) -> Result<Matrix, String>

Receives a list of SolarSurface objects as well as the Scene containing them and calculates the Daylight Coefficient Matrix that can be used for estimating the incident solar radiation in W/m2. The options for this calculation are contained in the DCFactory used as input.

pub fn make_fenestrations(
list: &[Rc<Fenestration>],
state: &mut SimulationStateHeader,
n_rays: usize
) -> Result<Vec<SolarSurface>, String>

Builds a set of SolarSurfaces from Fenestrations

Adds the necessary elements to the SimulationStateHeader

pub fn make_surfaces(
list: &[Rc<Surface>],
state: &mut SimulationStateHeader,
n_rays: usize
) -> Result<Vec<SolarSurface>, String>

Builds a set of SolarSurfaces from Surfaces

Adds the necessary elements to the `SimulationStateHeader

pub fn front_rays(&self) -> Vec<Ray3D>

Gets the front rays of a surface

pub fn back_rays(&self) -> Vec<Ray3D>

Gets the back rays of a surface

pub fn solar_irradiance(
&self,
rays: &[Ray3D],
scene: &Scene,
factory: &DCFactory
) -> Matrix

Calculates the Daylight Coefficient matrix for the front of a SolarSurface

pub fn calc_view_factors(
&self,
scene: &Scene,
front_side: bool
) -> Result<IRViewFactorSet, String>

Calculates an IRViewFactorSet for this surface